更新城市行业选择

+++ app/src/main/java/com/yushi/leke/fragment/exhibition/voteing/allproject/Industryinfo.java
@@ -1,13 +1,14 @@
 package com.yushi.leke.fragment.exhibition.voteing.allproject;
 
 public class Industryinfo {
-    private int industryId;
+    private long industryId;
     private String industryName;
-    public int getIndustryId() {
+
+    public long getIndustryId() {
         return industryId;
     }
 
-    public void setIndustryId(int industryId) {
+    public void setIndustryId(long industryId) {
         this.industryId = industryId;
     }
